idl - Moved our new namespaces to buildSrc.
